+++ WorldCities/WorldCities/src/app/auth/auth.interceptor.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
-import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpErrorResponse } from '@angular/common/http';
+import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpErrorResponse, HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
 import { catchError, Observable, throwError } from 'rxjs';
 import { AuthService } from './auth.service';
+import { isNullOrEmpty } from '../utils/Strings';
 
 @Injectable({
   providedIn: 'root'
@@ -20,7 +21,7 @@ export class AuthInterceptor implements HttpInterceptor {
   ): Observable<HttpEvent<any>>
   {
     // get the auth token
-    var token = this.authService.getToken();
+    var token = this.getProperToken();
 
     // if the token is present, clone the request
     // replacing the original headers with the authorization
@@ -36,11 +37,27 @@ export class AuthInterceptor implements HttpInterceptor {
     return next.handle(req)
       .pipe(catchError((error) => {
         // Perform Logout on 401 - Unauthorized HTTP response errors
-        if (error instanceof HttpErrorResponse && error.status === 401) {
+        if (error instanceof HttpErrorResponse && error.status === 401 && !this.authService.isAuthenticated()) {
           this.authService.logout();
           this.router.navigate(['login']);
         }
         return throwError(error);
       }));
   }
+
+
+  private getProperToken(): string | null
+  {
+    // Use this code block to determine what token to use and whether we need to refresh the tokens.
+    const curDateTime = new Date();
+    var IsUserAuth = this.authService.isAuthenticated();
+    var tokenToUse;
+    if (!IsUserAuth) return "";
+    var dateOfExpirationOnAccessToken = this.authService.getAccessTokenExp() as Date;
+    var IsAcccessTokenExpired = curDateTime >= dateOfExpirationOnAccessToken;
+    if (!IsAcccessTokenExpired) return this.authService.getAccessToken() as string;
+    this.authService.refreshToken();
+    return this.authService.getAccessToken();
+  }
+
 }