+++ WorldCities/WorldCities/src/app/utils/Strings.ts
@@ -1,0 +1,3 @@
+export function isNullOrEmpty(str: string | null | undefined): boolean {
+  return (str ?? '').trim().length === 0;
+}