+++ WorldCities/WorldCitiesAPI.Tests/CustomWebApplicationFactory.cs
@@ -1,19 +1,12 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+﻿
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.IdentityModel.Tokens;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Net;
 using WorldCitiesAPI.Data;
-using WorldCitiesAPI.Data.Models;
 
 namespace WorldCitiesAPI.Tests;
 
@@ -62,4 +55,5 @@ public class CustomWebApplicationFactory<TStartup> : WebApplicationFactory<TStar
       }
     });
   }
+
 }