+++ WorldCities/WorldCitiesAPI.Tests/HttpClientExtensions.cs
@@ -43,6 +43,7 @@ namespace WorldCitiesAPI.Tests
       if (httpResponse.IsSuccessStatusCode)
       {
         objResponse = JsonConvert.DeserializeObject<T>(await httpResponse.Content.ReadAsStringAsync());
+
       }
       else
       {