+++ WorldCities/WorldCitiesAPI/Controllers/AccountController.cs
@@ -1,9 +1,8 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
-using Azure;
-using Azure.Core;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using WorldCitiesAPI.Data;
 using WorldCitiesAPI.Data.Models;
 
@@ -22,7 +21,7 @@ namespace WorldCitiesAPI.Controllers
       ApplicationDbContext context,
       UserManager<ApplicationUser> userManager,
       SignInManager<ApplicationUser> signInManager,
-      JwtHandler jwtHandler) 
+      JwtHandler jwtHandler)
     {
       _context = context;
       _userManager = userManager;
@@ -33,7 +32,6 @@ namespace WorldCitiesAPI.Controllers
     [HttpPost("Register")]
     public async Task<IActionResult> Register([FromBody] RegisterRequest registerRequest)
     {
-
       // Checking if user via Username exists
       if (await _userManager.FindByNameAsync(registerRequest.UserName) != null)
       {
@@ -78,7 +76,7 @@ namespace WorldCitiesAPI.Controllers
       return Ok(new RegisterResult()
       {
         Success = true,
-        Message = "Registered User"
+        Message = "Registered Now. Please proceed to login."
       });
     }
 
@@ -115,14 +113,21 @@ namespace WorldCitiesAPI.Controllers
 
       var secToken = await _jwtHandler.GetAccessTokenAsync(user);
       var jwtAcc = new JwtSecurityTokenHandler().WriteToken(secToken);
-      response.Tokens.Add(jwtAcc);
+      response.Token = jwtAcc;
 
       var refToken = await _jwtHandler.GetRefreshTokenAsync(user);
       var jwtRef = new JwtSecurityTokenHandler().WriteToken(refToken);
-      response.Tokens.Add(jwtRef);
 
+      CookieOptions cookieOptions = new CookieOptions()
+      {
+        HttpOnly = true,
+        Secure = true,
+        SameSite = SameSiteMode.None,
+        Expires = DateTime.UtcNow.AddDays(_jwtHandler.GetRefreshTokenExpiration())
+      };
+
+      this.Response.Cookies.Append("refreshToken", jwtRef, cookieOptions);
       return Ok(response);
-
     }
 
 
@@ -130,9 +135,10 @@ namespace WorldCitiesAPI.Controllers
     [Authorize(Roles = "RegisteredUser")]
     public async Task<IActionResult> RefreshTokens([FromBody] ActiveUser activeUser)
     {
+      this.HttpContext.Request.Cookies.TryGetValue("refreshToken", out var refreshToken); // will add validation logic 
 
-      var user = await _userManager.FindByNameAsync(activeUser.Username);
-      if (user == null) return NotFound(new TokenRefresh() { NewToken = false, Message = "Member Not Found" });
+      var user = await _userManager.FindByEmailAsync(activeUser.Email);
+      if (user == null) return NotFound(new TokenRefresh() { NewToken = false, Message = "Email Not Found" });
 
       var response = new TokenRefresh()
       {
@@ -142,12 +148,22 @@ namespace WorldCitiesAPI.Controllers
 
       var secToken = await _jwtHandler.GetAccessTokenAsync(user);
       var jwtAcc = new JwtSecurityTokenHandler().WriteToken(secToken);
-      response.Tokens.Add(jwtAcc); // always add the access token first 
+      response.Token = jwtAcc; // always add the access token first 
 
       var refToken = await _jwtHandler.GetRefreshTokenAsync(user);
       var jwtRef = new JwtSecurityTokenHandler().WriteToken(refToken);
-      response.Tokens.Add(jwtRef);
 
+      CookieOptions cookieOptions = new CookieOptions()
+      {
+        HttpOnly = true,
+        IsEssential = true,
+        Secure = true,
+        SameSite = SameSiteMode.None,
+        Expires = DateTime.UtcNow.AddDays(_jwtHandler.GetRefreshTokenExpiration())
+      };
+
+      this.Response.Cookies.Delete("refreshToken");
+      this.Response.Cookies.Append("refreshToken", jwtRef, cookieOptions);
       return Ok(response);
     }