+++ WorldCities/WorldCitiesAPI/Data/ActiveUser.cs
@@ -2,6 +2,6 @@
 {
   public class ActiveUser
   {
-    public string Username { get; set; } = null!;
+    public string Email { get; set; } = null!;
   }
 }